+++ components/calendar/src/chinese.rs
@@ -665,7 +665,7 @@ mod test {
             iso_month: u8,
             iso_day: u8,
             expected_year: i32,
-            expected_month: u32,
+            expected_month: u8,
             expected_day: u32,
         }
 
@@ -1005,7 +1005,7 @@ mod test {
             iso_day: u8,
             expected_rel_iso: i32,
             expected_cyclic: u8,
-            expected_month: u32,
+            expected_month: u8,
             expected_day: u32,
         }
