+++ components/calendar/src/indian.rs
@@ -375,7 +375,7 @@ mod tests {
         iso_month: u8,
         iso_day: u8,
         expected_year: i32,
-        expected_month: u32,
+        expected_month: u8,
         expected_day: u32,
     }
